Refactor macros
The typescript is now generated with `mashin_bindgen` cli.

The macro now only generate the bindings.json which is used by the CLI to generate the final `mod.ts`

Now each resource can be implemented into their own `mod` and then the provider is built with a `construct_provider!` macro and then we can pass all resources to it.

This will make things much cleaner on big providers

Signed-off-by: david <david@lemarier.ca>