disable codesign (#13)
Gatekeeper only checks files which have been downloaded from the net using a browser, email or any other application, which sets the quarantine xattr flag on downloaded files. curl, scp, etc. do not know anything about this flag and thus files downloaded/copied with it are not quarantined.

Using the install script would fix it and doesnt require the binary to be signed for now